+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `custom_data_type_fixed_size` and `custom_data_type_variable_size` examples
 - Add `[Async]ArrayDlPackExt` traits that add methods to `Array` for `DLPack` tensor interop
   - Gated by the `dlpack` feature
+- Add missing `Group::async_child_*` methods
 
 ### Changed
 - **Breaking**: change `ArraySubset::inbounds` to take another subset rather than a shape
@@ -52,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Enables pass-through of codecs from Zarr V2 to V3 without converting to a V3 equivalent (if supported)
   - **Breaking**: Add `CodecTraits::{identifier,default_name,configuration[_opt]}()`
   - **Breaking**: Remove `CodecTraits::create_metadata[_opt]()`
+- **Breaking**: Change the error type of `node::[async_]get_child_nodes()` and `Group::{children,child_*}()` to `NodeCreateError` instead of `StorageError`
 
 ### Fixed
 - Fixed reserving one more element than necessary when retrieving `string` or `bytes` array elements